Time to think about an approach..
Found the source of the "collection mismatch" bug, turns out that
running a second thread in the background to modify the collection is
not going to work. Instead, the Application.Current.Dispatcher has to be
told to que up work that will modify the collection.

The problem now is, in order to scroll the ListBox after insertion of
new item the ViewModel needs a reference to the ListBox. But taking a
reference from the VIewModel to the view would make it harder to test
the ViewModel..

Need to now experiment with different approaches to see where this code
should go..
1) In the "code behind" file for the window.
2) In the xaml. Perhaps there is some parameter that I could set.